+++ libs/base_core/src/num/wide/mod.rs
@@ -1,0 +1,14 @@
+// devela_base_core::num::wide
+//
+#![doc = crate::_DOC_NUM_WIDE!()]
+//
+
+mod lane;
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            lane::*,
+        };
+    }
+}